Added Testfile fixed some Bugs
Added a simple file and one test for an example card (card: http://magiccards.info/m13/en/3.html) as well as some runtime text messages to make it easy to understand what is happening. Also added some show instances to data.
Fixed a bug not searching for Triggers on the Stack as it was added later.